feat: proxy media content in ODK Collect requests (#572)* refactor: split ODK Collect views* fix: cleaning* feat: proxy media content in ODK Collect requests
