+++ aether-odk-module/aether/odk/api/tests/test_models.py
@@ -84,7 +84,7 @@ class ModelsTests(CustomTestCase):
         self.assertEqual(instance.title, 'xForm - Test')
         self.assertEqual(instance.version, 'v1')
         self.assertEqual(instance.download_url,
-                         '/forms/{}/form.xml?version=v1'.format(instance.pk))
+                         f'/forms/{instance.pk}/form.xml?version=v1')
         self.assertEqual(instance.manifest_url, '', 'without media files no manifest url')
         self.assertEqual(str(instance), 'xForm - Test - xform-id-test')
 
@@ -113,13 +113,10 @@ class ModelsTests(CustomTestCase):
         self.assertEqual(media.md5sum, 'e2fc714c4727ee9395f324cd2e7f331f')
         self.assertEqual(media.hash, 'md5:e2fc714c4727ee9395f324cd2e7f331f')
         self.assertEqual(media.media_file_url, media.media_file.url)
-        self.assertEqual(
-            media.download_url,
-            f'{media.media_file_url}?response-content-disposition=attachment%3B%20filename%3D{media.name}'
-        )
+        self.assertEqual(media.download_url, f'/media-file/{media.pk}')
         # with media files there is manifest_url
         self.assertEqual(xform.manifest_url,
-                         '/forms/{}/manifest.xml?version={}'.format(xform.id, xform.version))
+                         f'/forms/{xform.id}/manifest.xml?version={xform.version}')
 
     def test__xform__version_control(self):
         xform = XForm.objects.create(
