+++ aether-odk-module/aether/odk/templates/xformManifest.xml
@@ -5,7 +5,7 @@
   <mediaFile>
     <filename>{{ media_file.name }}</filename>
     <hash>{{ media_file.hash }}</hash>
-    <downloadUrl>{{ media_file.download_url }}</downloadUrl>
+    <downloadUrl>{{ host }}{{ media_file.download_url }}</downloadUrl>
   </mediaFile>
   {% endfor %}
 </manifest>
